requirements.txt
pandas==1.4.4
psycopg2-binary==2.9.5
SQLAlchemy==1.4.45
